# src/generated/swagger.json
{
	"components": {
...
		"schemas": {
			"CustomerResponse": {
				"description": "Customer information response",
				"properties": {
					"id": {
						"type": "string",
						"description": "Unique customer identifier (UUID)"
					},
					"name": {
						"type": "string",
						"description": "Customer's full name"
					},
					"email": {
						"type": "string",
						"description": "Customer's email address"
					},
					"phone": {
						"type": "string",
						"description": "Phone number"
					},
					"defaultAddress": {
						"type": "string",
						"description": "Default delivery address"
					},
					"defaultLocation": {
						"properties": {
							"lng": {
								"type": "number",
								"format": "double"
							},
							"lat": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"lng",
							"lat"
						],
						"type": "object",
						"description": "Default location coordinates"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "Account creation timestamp"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"phone",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"id": "550e8400-e29b-41d4-a716-446655440000",
					"name": "Jane Smith",
					"email": "jane@example.com",
					"phone": "+1234567890",
					"defaultAddress": "123 Main St, City, State 12345",
					"createdAt": "2024-01-15T10:30:00.000Z"
				}
			},
			"CreateCustomerRequest": {
				"description": "Request body for creating a new customer (registration)",
				"properties": {
					"name": {
						"type": "string",
						"description": "Customer's full name"
					},
					"email": {
						"type": "string",
						"description": "Customer's email address (unique)"
					},
					"phone": {
						"type": "string",
						"description": "Phone number with country code (e.g., +1234567890, unique)"
					},
					"defaultAddress": {
						"type": "string",
						"description": "Default delivery address"
					},
					"defaultLocation": {
						"properties": {
							"lng": {
								"type": "number",
								"format": "double",
								"description": "Longitude coordinate"
							},
							"lat": {
								"type": "number",
								"format": "double",
								"description": "Latitude coordinate"
							}
						},
						"required": [
							"lng",
							"lat"
						],
						"type": "object",
						"description": "Default GPS location"
					}
				},
				"required": [
					"name",
					"email",
					"phone"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"name": "Jane Smith",
					"email": "jane@example.com",
					"phone": "+1234567890",
					"defaultAddress": "123 Main St, City, State 12345",
					"defaultLocation": {
						"lat": 37.7749,
						"lng": -122.4194
					}
				}
			},
			"UpdateCustomerRequest": {
				"description": "Request body for updating customer information",
				"properties": {
					"name": {
						"type": "string",
						"description": "Customer's full name"
					},
					"email": {
						"type": "string",
						"description": "Customer's email address"
					},
					"phone": {
						"type": "string",
						"description": "Phone number with country code"
					},
					"defaultAddress": {
						"type": "string",
						"description": "Default delivery address"
					},
					"defaultLocation": {
						"properties": {
							"lng": {
								"type": "number",
								"format": "double",
								"description": "Longitude coordinate"
							},
							"lat": {
								"type": "number",
								"format": "double",
								"description": "Latitude coordinate"
							}
						},
						"required": [
							"lng",
							"lat"
						],
						"type": "object",
						"description": "Default GPS location"
					}
				},
				"type": "object",
				"additionalProperties": false,
				"example": {
					"name": "Jane Smith Updated",
					"defaultAddress": "456 New St, City, State 54321"
				}
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "EdgeRun API",
		"version": "1.0.0",
		"description": "EdgeRun Last-Mile Delivery API - MVP for food delivery driver-order matching",
		"license": {
			"name": {
				"name": "MIT"
			}
		},
		"contact": {
			"name": "EdgeRun Support"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/customers": {
			"post": {
				"operationId": "RegisterCustomer",
				"responses": {
					"200": {
						"description": "Created customer information",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomerResponse"
								}
							}
						}
					},
					"201": {
						"description": "Customer created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomerResponse"
								}
							}
						}
					},
					"400": {
						"description": "Validation error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "Email or phone already registered",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Register a new customer",
				"tags": [
					"Customers"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Customer registration data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCustomerRequest",
								"description": "Customer registration data"
							}
						}
					}
				}
			}
		},
		"/customers/{id}": {
			"get": {
				"operationId": "GetCustomer",
				"responses": {
					"200": {
						"description": "Customer information",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomerResponse"
								}
							}
						}
					},
					"404": {
						"description": "Customer not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get customer by ID",
				"tags": [
					"Customers"
				],
				"security": [],
				"parameters": [
					{
						"description": "Customer ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCustomer",
				"responses": {
					"200": {
						"description": "Updated customer information",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomerResponse"
								}
							}
						}
					},
					"404": {
						"description": "Customer not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "Email or phone already registered",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Update customer information",
				"tags": [
					"Customers"
				],
				"security": [],
				"parameters": [
					{
						"description": "Customer ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Updated customer data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCustomerRequest",
								"description": "Updated customer data"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCustomer",
				"responses": {
					"200": {
						"description": "Success message",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Cannot delete customer with orders",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Customer not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Delete customer (admin only)",
				"tags": [
					"Customers"
				],
				"security": [],
				"parameters": [
					{
						"description": "Customer ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/customers/email/{email}": {
			"get": {
				"operationId": "GetCustomerByEmail",
				"responses": {
					"200": {
						"description": "Customer information",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomerResponse"
								}
							}
						}
					},
					"404": {
						"description": "Customer not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get customer by email",
				"tags": [
					"Customers"
				],
				"security": [],
				"parameters": [
					{
						"description": "Customer email address",
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}